+++ lib/repository/export_rides_repository.dart
@@ -1,5 +1,5 @@
 import 'package:weforza/database/export_rides_dao.dart';
-import 'package:weforza/model/exportable_ride.dart';
+import 'package:weforza/model/export/exportable_ride.dart';
 
 class ExportRidesRepository {
   ExportRidesRepository(this.dao);
